+++ rtapipe/scripts/ml/dataset_generation/single_long_photon_list/generate_for_testing.sh
@@ -10,7 +10,7 @@ echo "base_output_dir=$base_output_dir"
 output_dir_sim="$base_output_dir/sim_output"
 
 echo "Generating the photon list.."
-#simGRBcatalogWithRandomizationy -f config_test_set.yml --output-dir $output_dir_sim --print no --mp-threads 10 2>&1 > $this_dir/simulation_test.log
+#simGRBcatalogWithRandomization -f config_test_set.yml --output-dir $output_dir_sim --print no --mp-threads 10 2>&1 > $this_dir/simulation_test.log
 
 echo "Generating AP data.."
 data_dir="$base_output_dir/sim_output/run0406_ID000126"
@@ -19,17 +19,17 @@ ap_logs_output_dir="$base_output_dir/ap_logs_output"
 mkdir -p $ap_logs_output_dir
 
 #T=1 with-normalization
-#sbatch --output="$ap_logs_output_dir/t1_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -itype te -itime 1 -rr 0.2 -norm yes -out $ap_output_dir -proc 10"
+#sbatch --output="$ap_logs_output_dir/t1_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -off 0 -itype te -itime 1 -rr 0.2 -norm yes -out $ap_output_dir -proc 10"
 #T=1 no-normalization
-#sbatch --output="$ap_logs_output_dir/t1_no_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -itype te -itime 1 -rr 0.2 -norm no -out $ap_output_dir -proc 10"
+#sbatch --output="$ap_logs_output_dir/t1_no_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -off 0 -itype te -itime 1 -rr 0.2 -norm no -out $ap_output_dir -proc 10"
 #T=5 with-normalization
-#sbatch --output="$ap_logs_output_dir/t5_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -itype te -itime 5 -rr 0.2 -norm yes -out $ap_output_dir -proc 10"
+#sbatch --output="$ap_logs_output_dir/t5_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -off 0 -itype te -itime 5 -rr 0.2 -norm yes -out $ap_output_dir -proc 10"
 #T=5 no-normalization
-#sbatch --output="$ap_logs_output_dir/t5_no_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -itype te -itime 5 -rr 0.2 -norm no -out $ap_output_dir -proc 10"
+#sbatch --output="$ap_logs_output_dir/t5_no_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -off 0 -itype te -itime 5 -rr 0.2 -norm no -out $ap_output_dir -proc 10"
 #T=10 with-normalization
-#sbatch --output="$ap_logs_output_dir/t10_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -itype te -itime 10 -rr 0.2 -norm yes -out $ap_output_dir -proc 10"
+#sbatch --output="$ap_logs_output_dir/t10_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -off 0 -itype te -itime 10 -rr 0.2 -norm yes -out $ap_output_dir -proc 10"
 #T=10 no-normalization
-#sbatch --output="$ap_logs_output_dir/t10_no_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -itype te -itime 10 -rr 0.2 -norm no -out $ap_output_dir -proc 10"
+#sbatch --output="$ap_logs_output_dir/t10_no_norm.log" --wrap="generate_ap_data -c $this_dir/config_test_set.yml -dd $data_dir -off 0 -itype te -itime 10 -rr 0.2 -norm no -out $ap_output_dir -proc 10"
 
 
 echo "Generating plots.."